+++ src/model/mod.rs
@@ -1,5 +1,7 @@
 mod geometry;
+mod preset;
 mod tree;
 
 pub use geometry::*;
+pub use preset::*;
 pub use tree::*;